Complete query authority cutover and Worth UI integration

+++ crates/worth-query/tests/ui/typed_contains_requires_string_field.rs
@@ -3,7 +3,7 @@ use worth_query::facade::runtime::TypedDetailQueryBuilder;
 worth_query::worth_query_schema! {
     pub schema UserSchema("user") {
         fields {
-            pub field Age("profile", "age", Integer) => [projectable, equality(i64), orderable];
+            pub field Age("profile", "age", Int64) => [projectable, equality(i64), orderable];
         }
         relations {}
     }
